Final TOuches-> Working on database design

+++ app/Http/Controllers/Admin/CategoryController.php
@@ -5,6 +5,7 @@ namespace App\Http\Controllers\Admin;
 use App\DataTables\CategoryDataTable;
 use App\Http\Controllers\Controller;
 use App\Models\Category;
+use App\Models\PortfolioItem;
 use Illuminate\Http\Request;
 
 class CategoryController extends Controller
@@ -92,6 +93,12 @@ class CategoryController extends Controller
     {
         //
         $category = Category::findorfail($id);
-        $category->delete();
+        $hasItem = PortfolioItem::where('category_id', $category->id)->count();
+        if($hasItem == 0)
+        {
+            $category->delete();
+            return true;
+        }
+        return response(['status' => 'error']);
     }
 }
